requirements.txt
biopython==1.85
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.3
jax==0.4.20
jaxlib==0.4.20
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
ml-dtypes==0.3.1
numpy==1.26.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
